Filter out scoped packages
The npm endpoint used to retrieve download counts doesn't work with scoped packages (or at least not in a way that I can find), and causes the `nud` command to fail if applied to a user who has scoped packages.

Fix this by filtering scoped packages out of the package list before passing to `npm-package-downloads`.